requirements.txt
djangorestframework~=3.15.2
Django~=5.1.1
djoser~=2.2.3
python-dotenv~=1.0.1
django_bootstrap5~=24.2
flake8=7.1.1